Updated policy on EOS frequency
It's already the case but we want this in policy. So instead of idling by waiting for a support ticket to come, you can resume to doing your regular tasks once you've "cleaned" GrooveHQ. "Cleaned" can either be zero emails in Groove or no more open tickets that are assigned to you.