Clarify the conference fee reimbursement
The fees for conferences can be quite high and the conference allowance would not cover fees + flight + hotel. Therefore after quick discussion with Dejan we had the idea to change the conference fee to be a business expense.

The wording has been updated to refelect this and the example trip calculations updated.

Additional US spelling and grammar corrections.