Productivity Allowance budget is given in the same currency as your salary
When explaining to @dmurko the recent changes in Scrooge we again, for the
N-th time got stuck with how allowance is calculated. His understanding was
that [the previous PR](https://github.com/niteoweb/handbook/pull/135) only
applied to those that receive their salary in EUR.

A long debate ensued and we came out with a realization that everyone should
receive their allowance budget in the currency they receive their salary in.

For example: I'm on a conference and want to buy dinner for an author of an
Open Source library that we use a lot as a thank-you gesture. I remember that
I have $20 left of allowance. Uhm, what is the current exchange rate to EUR?
Ugh, I need to fire up my phone and check before I feel comfortable spending
the allowance. Now imagine my allowance is always in my "base" currency. If
the amount is 20€, that is not gonna change in value on me all of a sudden.

BTW, this decision was one of the reasons why [Scrooge needed a rewrite](
https://github.com/niteoweb/scrooge/issues/24).