Introduce Alignment Meetings
During the discussion on https://github.com/niteoweb/operations/issues/457
we realized that our Retrospetives are too long and some Niteans are getting
bored during them.

We propose instead to make the Retrospective shorter and only about the
scrum process. Then when we are done with the retrospective, only those who
are interested gather again to discus more general topics such as reviewing spreadsheets and charts. The rest of Niteans can get back to their work.