Revert "Add instructions on how to add story estimates."
We really did fail our Scrum process here. Changes to Scrum process need
to be discussed/approved on Retrospectives so that everyone knows about them.

I should not have merged https://github.com/niteoweb/handbook/pull/140. It was
great that we had the PR, but it should not have been merged before the
Retrospective.

Ideal process is:
* a Nitean has an idea to change how we do Scrum
* they create a PR with the proposal
* whoever is interested polishes the idea during the sprint
* idea is presented during Retrospective and the PR is merged if majority
  approves it.

This reverts commit a9a82a5e51ecc53fc5f43049b96371f5ee0431f0.