Update with correct meeting timezone
Clarification was required for the morning standup meetings and the
reasoning needed noting down.

Daily standup page updated with correct meeting time of 10 AM CET. In
addition a paragraph about timezone and how we came to the decision of
the morning meeting time.

References to the specific standup meeting time were removed from all
pages, replaced with link to standup.md page. This is so that the
meeting time/timezone need only be changed in one location in future.

Cleanup spelling, grammar and other small details.