Am I allowed to make my business trip into vacation trip? (#156)
@yevnik asked today: "There's no flight with a reasonable layover to
Slovenia on 8th of July, so I'm thinking to fly through Italy. Would
you mind if I book the flight earlier, spend several vacation days in
Italy and get to Ljubljana by bus?"

This commit is the answer. Both me and @dmurko have done this in the past
numerous times, @cas-- did it after IRL#4 in Istanbul, etc. It really does
make sense that if you are already at some new place to take advantage of
the situation.