Remove details about the vote prefix
As part of the change to using the Estimation pipeline it was decided
that we no longer have any need for the vote prefix since the majority
of user stories submitted for review and voting are ready to be voted
upon. It also saves time by not having to apply a vote prefix to the
issue title.

Refs: niteoweb/operations#572