Remove "business expense" (#178)
* Remove "business expense" 

It was just causing confusions all the time. Also, what if a conference fee includes accommodation and food and drinks? Was it then business expense or allowance?

Now we have just allowance. And everyone that travels regularly gets an additional €1000 bonus every January to encourage us to go to conferences and events.

* Every Nitean should start with €1000 not €0

So they can afford to go to a conf/IRL right of the bat, buy a better headset, order a faster internet connection.

* Typos