Update overtime system to reward vacation hours
As discussed on [IRL #FML](teamniteo/operations#1586 (comment)) we will, for half a year or so, try to have overtime
be rewarded with 1.3 times the amount in vacation hours instead of overtime pay.
This new system should allow all Niteans to work the way that feels best for them without potentially affecting e.g. profit sharing.

Ref: https://github.com/teamniteo/operations/issues/1739